Moved test attributes in dedicated folder
