+++ src/AoPeas.Tests/Fixtures/Sut/NoImplementedPointcut.cs
@@ -1,0 +1,3 @@
+namespace AoPeas.Tests.Fixtures.Sut;
+
+public class NoImplementedPointcutAttribute : PointcutAttribute { }
