+++ src/AoPeas.Tests/Fixtures/Sut/PassthroughPointcut.cs
@@ -1,0 +1,8 @@
+namespace AoPeas.Tests.Fixtures.Sut;
+
+public class PassthroughPointcutAttribute : PointcutAttribute { }
+
+public class PassthroughPointcut : IAdvice<PassthroughPointcutAttribute>
+{
+    public object? Apply(MethodInvocationDetails invocationDetails) => invocationDetails.Next();
+}
