+++ src/AoPeas.Tests/Fixtures/TestService.cs
@@ -1,5 +1,4 @@
-using AoPeas;
-using Microsoft.Extensions.Logging;
+using AoPeas.Tests.Fixtures.Sut;
 
 namespace AoPeas.Tests.Fixtures;
 
@@ -12,63 +11,10 @@ public interface ITestService
 
 public class TestService : ITestService
 {
-    [NoAdvice]
-    [AddTen]
+    [PassthroughPointcut]
     public int GetIncrement(int a) => a + 1;
-    [NotImplementedPointcut]
+    [NoImplementedPointcut]
     public int GetSum(int a, int b) => a + b;
-    [EnableSecondProxyLogging]
-    [EnableProxyLogging]
+    [LogCalls]
     public int GetSum(int a, int b, int c) => a + b + c;
-}
-
-public class NotImplementedPointcutAttribute : PointcutAttribute { }
-
-public class NoAdviceAttribute : PointcutAttribute { }
-
-public class NoAdvice : IAdvice<NoAdviceAttribute>
-{
-    public object? Apply(MethodInvocationDetails invocationDetails) => invocationDetails.Next();
-}
-
-public class AddTenAttribute : PointcutAttribute { }
-
-public class AddTen : IAdvice<AddTenAttribute>
-{
-    public object? Apply(MethodInvocationDetails invocationDetails) => (invocationDetails.Next() as int?) + 10;
-}
-
-public class EnableProxyLoggingAttribute : PointcutAttribute { }
-
-public class LoggingAdvice(ILogger<LoggingAdvice> logger) : IAdvice<EnableProxyLoggingAttribute>
-{
-    private int count = 0;
-
-    public object? Apply(MethodInvocationDetails invocationDetails)
-    {
-        count++;
-        logger.LogInformation("Start method '{MethodInfo}', arguments: {Arguments}, count: {count}", invocationDetails.Name, string.Join(',', invocationDetails.Args), count);
-
-        var result = invocationDetails.Next();
-
-        logger.LogInformation("End method '{MethodInfo}', result: {Result}", invocationDetails.Name, result);
-
-        return result;
-    }
-}
-
-public class EnableSecondProxyLoggingAttribute : PointcutAttribute { }
-
-public class SecondLoggingAdvice(ILogger<SecondLoggingAdvice> logger) : IAdvice<EnableSecondProxyLoggingAttribute>
-{
-    public object? Apply(MethodInvocationDetails invocationDetails)
-    {
-        logger.LogInformation("Start method '{MethodInfo}', arguments: {Arguments}", invocationDetails.Name, string.Join(',', invocationDetails.Args));
-
-        var result = invocationDetails.Next();
-
-        logger.LogInformation("End method '{MethodInfo}', result: {Result}", invocationDetails.Name, result);
-
-        return result;
-    }
 }
