+++ src/AoPeas.Tests/Tests/AopProxyTests.cs
@@ -1,5 +1,6 @@
 using AoPeas.Internal;
 using AoPeas.Tests.Fixtures;
+using AoPeas.Tests.Fixtures.Sut;
 
 namespace AoPeas.Tests.Tests;
 
@@ -9,21 +10,10 @@ public class AopProxyTests
     public void GivenAdvice_WhenCallingMethod_ReturnsCorrectResult()
     {
         var testService = new TestService();
-        var aspectMap = new AspectMap(new() { [typeof(NoAdviceAttribute)] = [new NoAdvice()] });
+        var aspectMap = new AspectMap(new() { [typeof(PassthroughPointcutAttribute)] = [new PassthroughPointcut()] });
 
         var proxy = (ITestService)AopProxy.Create(typeof(ITestService), testService, aspectMap);
 
         Assert.Equal(4, proxy.GetIncrement(3));
-    }
-
-    [Fact]
-    public void GivenAdvice_WhenAdjustingResult_ReturnsModifiedResult()
-    {
-        var testService = new TestService();
-        var aspectMap = new AspectMap(new() { [typeof(AddTenAttribute)] = [new AddTen()] });
-
-        var proxy = (ITestService)AopProxy.Create(typeof(ITestService), testService, aspectMap);
-
-        Assert.Equal(14, proxy.GetIncrement(3));
     }
 }
